+++ enterprise/storage/saas_secrets_store.py
@@ -13,7 +13,7 @@ from storage.user_store import UserStore
 from openhands.core.config.openhands_config import OpenHandsConfig
 from openhands.core.logger import openhands_logger as logger
 from openhands.storage.data_models.secrets import Secrets
-from openhands.storage.secrets.secrets_store import SecretsStore
+from openhands.app_server.secrets.secrets_store import SecretsStore
 
 
 @dataclass
