+++ openhands/server/shared.py
@@ -11,6 +11,7 @@ import os
 import socketio
 from dotenv import load_dotenv
 
+from openhands.app_server.secrets.secrets_store import SecretsStore
 from openhands.app_server.settings.settings_store import SettingsStore
 from openhands.core.config import load_openhands_config
 from openhands.core.config.openhands_config import OpenHandsConfig
@@ -18,7 +19,6 @@ from openhands.server.config.server_config import ServerConfig, load_server_conf
 from openhands.server.types import ServerConfigInterface
 from openhands.storage import get_file_store
 from openhands.storage.files import FileStore
-from openhands.storage.secrets.secrets_store import SecretsStore
 from openhands.utils.import_utils import get_impl
 
 load_dotenv()
