+++ openhands/server/user_auth/__init__.py
@@ -2,12 +2,12 @@ from fastapi import Depends, Request
 from fastapi.security import APIKeyHeader
 from pydantic import SecretStr
 
+from openhands.app_server.secrets.secrets_store import SecretsStore
 from openhands.app_server.settings.settings_store import SettingsStore
 from openhands.integrations.provider import PROVIDER_TOKEN_TYPE
 from openhands.server.settings import Settings
 from openhands.server.user_auth.user_auth import AuthType, get_user_auth
 from openhands.storage.data_models.secrets import Secrets
-from openhands.storage.secrets.secrets_store import SecretsStore
 
 
 async def get_provider_tokens(request: Request) -> PROVIDER_TOKEN_TYPE | None:
