+++ tests/unit/app_server/test_secrets_api.py
@@ -9,6 +9,7 @@ from fastapi import FastAPI
 from fastapi.testclient import TestClient
 from pydantic import SecretStr
 
+from openhands.app_server.secrets.file_secrets_store import FileSecretsStore
 from openhands.app_server.secrets.secrets_router import (
     router as secrets_router,
 )
@@ -19,7 +20,6 @@ from openhands.integrations.provider import (
 )
 from openhands.storage import get_file_store
 from openhands.storage.data_models.secrets import Secrets
-from openhands.storage.secrets.file_secrets_store import FileSecretsStore
 
 
 @pytest.fixture
